+++ libs/utils/include/celix_string_hash_map.h
@@ -23,6 +23,7 @@
 #include "celix_cleanup.h"
 #include "celix_hash_map_value.h"
 #include "celix_utils_export.h"
+#include "celix_errno.h"
 
 /**
  * @brief Init macro so that the opts are correctly initialized for C++ compilers
@@ -202,44 +203,53 @@ CELIX_UTILS_EXPORT size_t celix_stringHashMap_size(const celix_string_hash_map_t
 /**
  * @brief Add pointer entry the string hash map.
  *
- * @note The returned previous value can be already freed by a removed callback (if configured).
+ * If the return status is an error, a error message is logged to celix_err.
  *
  * @param[in] map The hashmap.
  * @param[in] key  The key to use. The hashmap will create a copy if needed.
  * @param[in] value The value to store with the key.
- * @return The previous value or NULL of no value was set for th provided key.
- *         Note also returns NULL if the previous value for the key was NULL.
+ * @return celix_status_t CELIX_SUCCESS if the entry was added, CELIX_ENOMEM if no memory could be allocated for the
+ *         entry.
  */
-CELIX_UTILS_EXPORT void* celix_stringHashMap_put(celix_string_hash_map_t* map, const char* key, void* value);
+CELIX_UTILS_EXPORT celix_status_t celix_stringHashMap_put(celix_string_hash_map_t* map, const char* key, void* value);
 
 /**
  * @brief add long entry the string hash map.
  *
+ * If the return status is an error, a error message is logged to celix_err.
+ *
  * @param map The hashmap.
  * @param key  The key to use. The hashmap will create a copy if needed.
  * @param value The value to store with the key.
- * @return True if a previous value with the provided has been replaced.
+ * @return celix_status_t CELIX_SUCCESS if the entry was added, CELIX_ENOMEM if no memory could be allocated for the
+ *         entry.
  */
-CELIX_UTILS_EXPORT bool celix_stringHashMap_putLong(celix_string_hash_map_t* map, const char* key, long value);
+CELIX_UTILS_EXPORT celix_status_t celix_stringHashMap_putLong(celix_string_hash_map_t* map, const char* key, long value);
 
 /**
  * @brief add double entry the string hash map.
  *
+ * If the return status is an error, a error message is logged to celix_err.
+ *
  * @param map The hashmap.
  * @param key  The key to use. The hashmap will create a copy if needed.
- * @return True if a previous value with the provided has been replaced.
+ * @return celix_status_t CELIX_SUCCESS if the entry was added, CELIX_ENOMEM if no memory could be allocated for the
+ *         entry.
  */
-CELIX_UTILS_EXPORT bool celix_stringHashMap_putDouble(celix_string_hash_map_t* map, const char* key, double value);
+CELIX_UTILS_EXPORT celix_status_t celix_stringHashMap_putDouble(celix_string_hash_map_t* map, const char* key, double value);
 
 /**
  * @brief add bool entry the string hash map.
  *
+ * If the return status is an error, a error message is logged to celix_err.
+ *
  * @param map The hashmap.
  * @param key  The key to use. The hashmap will create a copy if needed.
  * @param value The value to store with the key.
- * @return True if a previous value with the provided has been replaced.
+ * @return celix_status_t CELIX_SUCCESS if the entry was added, CELIX_ENOMEM if no memory could be allocated for the
+ *         entry.
  */
-CELIX_UTILS_EXPORT bool celix_stringHashMap_putBool(celix_string_hash_map_t* map, const char* key, bool value);
+CELIX_UTILS_EXPORT celix_status_t celix_stringHashMap_putBool(celix_string_hash_map_t* map, const char* key, bool value);
 
 /**
  * @brief Returns the value for the provided key.